--- src/pactTest/java/uk/gov/justice/laa/amend/claim/config/ClaimsApiPactTestConfig.java
@@ -1,6 +1,0 @@
-package uk.gov.justice.laa.amend.claim.config;
-
-import org.springframework.boot.test.context.TestConfiguration;
-
-@TestConfiguration
-public class ClaimsApiPactTestConfig {}
