--- src/pactTest/resources/pact.properties
@@ -1,3 +1,0 @@
-pactbroker.url = http://localhost:9292
-pactbroker.auth.username = $PACT_USER
-pactbroker.auth.password = $PACT_PASSWORD
